Share components between the blog post and company update post
