+++ src/templates/blog-post/index.tsx
@@ -1,401 +1,36 @@
-import { graphql, Link } from 'gatsby';
+import { graphql } from 'gatsby';
 import * as React from 'react';
-
 import dayjs from 'dayjs';
 import reltime from 'dayjs/plugin/relativeTime';
-
-import CalendarTodayOutlined from '@mui/icons-material/CalendarTodayOutlined';
-import DoneIcon from '@mui/icons-material/Done';
-import { GatsbyImage, getImage } from 'gatsby-plugin-image';
-import { Divider, useTheme } from '@mui/material';
-import SwoopingLinesBackground from '../../components/BackgroundImages/LightSwoopingLinesRightDirectionBackground';
-import StraightLinesBackground from '../../components/BackgroundImages/StraightLinesBackground';
-import { PopularArticles } from '../../components/BlogPopularArticles';
 import BlogPostPopupModal from '../../components/BlogPostPopupModal';
-import { ProcessedPost } from '../../components/BlogPostProcessor';
 import Breadcrumbs from '../../components/Breadcrumbs';
-import Bio from '../../components/Bio';
 import Layout from '../../components/Layout';
 import Seo from '../../components/seo';
 import logoUrl from '../../images/estuary.png';
-import { costPerGB } from '../../utils';
-import ReadingTimeIcon from '../../svgs/time.svg';
-import {
-    dashboardRegisterUrl,
-    getAuthorPathBySlug,
-    getAuthorSeoJson,
-} from '../../../shared';
-import Avatar from '../../components/Avatar';
-import SocialLinks from '../../components/SocialLinks';
-import BlogBanner from '../../components/BlogBanner';
-import ArticleSidebar from '../../components/ArticleSidebar';
-import Container from '../../components/Container';
-import HeroSectionDetails from '../../components/HeroSectionDetails';
-import OutboundLinkFilled from '../../components/LinksAndButtons/OutboundLinkFilled';
-import OutboundLinkOutlined from '../../components/LinksAndButtons/OutboundLinkOutlined';
-import ShareArticle from '../../components/ShareArticle';
-import {
-    blogPost,
-    blogPostHeaderWrapper,
-    headerInfo,
-    postInfo,
-    tagsWrapper,
-    blogsPostCardTags,
-    dateAndReadWrapper,
-    iconInfoWrapper,
-    blogPostDate,
-    heroImage,
-    shareArticleMobile,
-    blogPostContent,
-    blogPostContentWrapper,
-    mainContent,
-    bigBuildPipelineBannerContainer,
-    popularArticlesWrapper,
-    bigBuildPipelineBannerSection,
-    bigBuildPipelineBannerWrapper,
-    bigBuildPipelineBannerContainerLayout,
-    leftColumnContainer,
-    rightColumnContainer,
-    buildPipelineAndPricingButtons,
-    pricingLink,
-    blogPostBreadcrumbsWrapper,
-    nextStepsAndAboutAuthorSection,
-    aboutAuthor,
-    authorInfo,
-    authorMainInfoContainer,
-    authorAvatarContainer,
-    authorNameAndRole,
-    authorName,
-    authorRole,
-} from './styles.module.less';
+import { getAuthorSeoJson } from '../../../shared';
+import BlogPost from '../../components/BlogPost';
 
 dayjs.extend(reltime);
 
 const BlogPostTemplate = ({ data: { post } }) => {
-    const theme = useTheme();
-
-    const postTags = post?.tags?.filter((tag) => tag.type === 'tag');
-
-    const hasBeenUpdated = post?.updatedAt
-        ? post?.publishedAt !== post?.updatedAt
-        : false;
-
-    const hasAtLeastOneBio = post?.authors.some(
-        (author) => author?.bio?.data?.bio
-    );
-    const tableOfContents = post.body.data.childHtmlRehype.tableOfContents;
-
     return (
         <Layout>
-            <div className={blogPostBreadcrumbsWrapper}>
-                <Breadcrumbs
-                    breadcrumbs={[
-                        {
-                            title: 'Home',
-                            href: '/',
-                        },
-                        {
-                            title: 'Blog',
-                            href: '/blog',
-                        },
-                        {
-                            title: post.title,
-                        },
-                    ]}
-                />
-            </div>
-            <article
-                className={blogPost}
-                itemScope
-                itemType="http://schema.org/Article"
-            >
-                <SwoopingLinesBackground>
-                    <Container className={blogPostHeaderWrapper}>
-                        <div className={headerInfo}>
-                            <div className={postInfo}>
-                                <div className={tagsWrapper}>
-                                    {postTags.map((tag) => (
-                                        <span
-                                            key={tag.name}
-                                            className={blogsPostCardTags}
-                                        >
-                                            {tag.name}
-                                        </span>
-                                    ))}
-                                </div>
-                                <div className={dateAndReadWrapper}>
-                                    <div className={iconInfoWrapper}>
-                                        <ReadingTimeIcon color="#47506D" />
-                                        <span className={blogPostDate}>
-                                            {
-                                                post.body.data
-                                                    .childMarkdownRemark.fields
-                                                    .readingTime.text
-                                            }
-                                        </span>
-                                    </div>
-                                    <div className={iconInfoWrapper}>
-                                        <CalendarTodayOutlined fontSize="small" />
-                                        <span className={blogPostDate}>
-                                            <span>
-                                                {hasBeenUpdated
-                                                    ? `Published ${post.publishedAt}`
-                                                    : post.publishedAt}
-                                            </span>
-                                            {hasBeenUpdated ? (
-                                                <span>
-                                                    Updated {post.updatedAt}
-                                                </span>
-                                            ) : null}
-                                        </span>
-                                    </div>
-                                </div>
-                            </div>
-                            <HeroSectionDetails
-                                title={post.title}
-                                description={post.description}
-                            />
-                            <Bio authors={post.authors} />
-                        </div>
-                        {post.hero ? (
-                            <GatsbyImage
-                                alt={post.title}
-                                className={heroImage}
-                                image={
-                                    post.hero.localFile.childImageSharp
-                                        .gatsbyImageData
-                                }
-                                loading="eager"
-                            />
-                        ) : null}
-                        <div className={shareArticleMobile}>
-                            <ShareArticle
-                                article={{
-                                    title: post.title,
-                                    slug: post.slug,
-                                }}
-                            />
-                        </div>
-                    </Container>
-                </SwoopingLinesBackground>
-
-                {post.body ? (
-                    <section className={blogPostContent}>
-                        <div className={blogPostContentWrapper}>
-                            <div className={mainContent}>
-                                <ProcessedPost
-                                    body={post.body.data.childHtmlRehype.html}
-                                />
-                                <BlogBanner
-                                    title={
-                                        <h3>
-                                            Start streaming your data{' '}
-                                            <span>for free</span>
-                                        </h3>
-                                    }
-                                    button={{
-                                        title: 'Build a Pipeline',
-                                        href: dashboardRegisterUrl,
-                                    }}
-                                />
-                            </div>
-                            <ArticleSidebar
-                                article={{
-                                    title: post.title,
-                                    slug: post.slug,
-                                }}
-                                tableOfContents={tableOfContents}
-                            />
-                        </div>
-                    </section>
-                ) : null}
-                {post?.authors.length >= 1 ? (
-                    <section className={nextStepsAndAboutAuthorSection}>
-                        {/* <div className={nextSteps}>
-                        <h3>Next steps</h3>
-                        <NextStepsLink href="">Read about Lorem ipsum dolor sit amet, consectetur</NextStepsLink>
-                        <NextStepsLink href="">Learn about Lorem ipsum dolor sit amet</NextStepsLink>
-                        <NextStepsLink href="">Lorem ipsum dolor sit amet</NextStepsLink>
-                        </div> */}
-
-                        <div className={aboutAuthor}>
-                            <h2>
-                                {post?.authors.length === 1
-                                    ? hasAtLeastOneBio
-                                        ? 'About the author'
-                                        : 'Author'
-                                    : hasAtLeastOneBio
-                                      ? 'About the authors'
-                                      : 'Authors'}
-                            </h2>
-                            {post?.authors?.map((author, index) => {
-                                const authorImage =
-                                    author?.picture &&
-                                    getImage(
-                                        author.picture.localFile.childImageSharp
-                                            .gatsbyImageData
-                                    );
-
-                                const authorBio = author?.bio.data.bio;
-
-                                const authorSocialLinks = author?.socials;
-
-                                return (
-                                    <>
-                                        <div key={index} className={authorInfo}>
-                                            <Link
-                                                to={getAuthorPathBySlug(
-                                                    author?.slug
-                                                )}
-                                                className={
-                                                    authorMainInfoContainer
-                                                }
-                                            >
-                                                <div
-                                                    className={
-                                                        authorAvatarContainer
-                                                    }
-                                                >
-                                                    <Avatar
-                                                        image={authorImage}
-                                                        alt={`Picture of ${author?.name}`}
-                                                        name={author.name}
-                                                        loading="lazy"
-                                                        size={60}
-                                                    />
-                                                </div>
-                                                <div
-                                                    className={
-                                                        authorNameAndRole
-                                                    }
-                                                >
-                                                    {author?.name ? (
-                                                        <span
-                                                            className={
-                                                                authorName
-                                                            }
-                                                        >
-                                                            {author.name}
-                                                        </span>
-                                                    ) : null}
-                                                    {author?.role ? (
-                                                        <span
-                                                            className={
-                                                                authorRole
-                                                            }
-                                                        >
-                                                            {author.role}
-                                                        </span>
-                                                    ) : null}
-                                                </div>
-                                            </Link>
-                                            {authorSocialLinks ? (
-                                                <>
-                                                    <Divider
-                                                        orientation="vertical"
-                                                        variant="middle"
-                                                        flexItem
-                                                        sx={{
-                                                            minHeight: '57px',
-                                                            width: '1px',
-                                                            borderColor:
-                                                                '#d7dce5',
-                                                            margin: '0 30px 0 20px',
-                                                            [theme.breakpoints.down(
-                                                                520
-                                                            )]: {
-                                                                display: 'none',
-                                                            },
-                                                        }}
-                                                    />
-                                                    <SocialLinks
-                                                        socialLinks={
-                                                            authorSocialLinks
-                                                        }
-                                                    />
-                                                </>
-                                            ) : null}
-                                        </div>
-                                        {authorBio ? (
-                                            <div
-                                                dangerouslySetInnerHTML={{
-                                                    __html: author.bio.data.bio,
-                                                }}
-                                            />
-                                        ) : null}
-                                    </>
-                                );
-                            })}
-                        </div>
-                    </section>
-                ) : null}
-                <section className={popularArticlesWrapper}>
-                    <h2>Popular Articles</h2>
-                    <PopularArticles />
-                </section>
-                <section className={bigBuildPipelineBannerSection}>
-                    <div className={bigBuildPipelineBannerWrapper}>
-                        <StraightLinesBackground
-                            className={bigBuildPipelineBannerContainer}
-                        >
-                            <div
-                                className={
-                                    bigBuildPipelineBannerContainerLayout
-                                }
-                            >
-                                <div className={leftColumnContainer}>
-                                    <h5>Streaming Pipelines.</h5>
-                                    <h5>Simple to Deploy.</h5>
-                                    <h5>Simply Priced.</h5>
-                                </div>
-                                <div className={rightColumnContainer}>
-                                    <div>
-                                        <DoneIcon
-                                            htmlColor="#5072eb"
-                                            fontSize="large"
-                                        />
-                                        <span>
-                                            {costPerGB} of data moved +
-                                            $.14/connector/hour;
-                                        </span>
-                                    </div>
-                                    <div>
-                                        <DoneIcon
-                                            htmlColor="#5072eb"
-                                            fontSize="large"
-                                        />
-                                        <span>
-                                            50% less than competing ETL/ELT
-                                            solutions;
-                                        </span>
-                                    </div>
-                                    <div>
-                                        <DoneIcon
-                                            htmlColor="#5072eb"
-                                            fontSize="large"
-                                        />
-                                        <span>
-                                            {'<'}100ms latency on streaming
-                                            sinks/sources.
-                                        </span>
-                                    </div>
-                                </div>
-                            </div>
-                            <div className={buildPipelineAndPricingButtons}>
-                                <OutboundLinkFilled href={dashboardRegisterUrl}>
-                                    Build a Pipeline
-                                </OutboundLinkFilled>
-                                <OutboundLinkOutlined
-                                    href="https://estuary.dev/pricing/"
-                                    className={pricingLink}
-                                >
-                                    See Pricing
-                                </OutboundLinkOutlined>
-                            </div>
-                        </StraightLinesBackground>
-                    </div>
-                </section>
-            </article>
+            <Breadcrumbs
+                breadcrumbs={[
+                    {
+                        title: 'Home',
+                        href: '/',
+                    },
+                    {
+                        title: 'Blog',
+                        href: '/blog',
+                    },
+                    {
+                        title: post.title,
+                    },
+                ]}
+            />
+            <BlogPost post={post} />
             <BlogPostPopupModal />
         </Layout>
     );
