+++ src/templates/connector/styles.module.less
@@ -7,16 +7,3 @@
     min-width: 24px;
     min-height: 24px;
 }
-
-.breadcrumbsWrapper {
-    margin: auto;
-    padding: 40px 0;
-
-    @media (max-width: 810px) {
-        padding: 40px 0 20px;
-    }
-
-    @media (max-width: 768px) {
-        display: none;
-    }
-}
